+++ CHANGELOG.md
@@ -4,3 +4,4 @@
   `rollun\utils\Cleaner`, which has been removed in the latest version of 'rollun-com/rollun-utils'.
 - Removed rollun-com/rollun-installer as it was unused
 - Removed `rollun\files` as it was moved to another package
+- Make "count(): int" and "getIterator(): \Traversable" functions typed in DataStoreAbstract